Move repositories package to data package
