+++ app/src/main/java/com/shapeapp/shape/ui/messages/MessagesViewModel.kt
@@ -3,7 +3,7 @@ package com.shapeapp.shape.ui.messages
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import com.shapeapp.shape.data.entities.Message
-import com.shapeapp.shape.repositories.MessagesRepository
+import com.shapeapp.shape.data.repositories.MessagesRepository
 
 /**
  * [ViewModel] for [MessagesFragment]
